requirements.txt
lxml~=4.9.1
requests-html~=0.10.0

options==0.451
selenium==4.3.*
webdriver-manager==3.8.*